Updating
The download function has been added where it verifies that the file is being downloaded and is informed when the download is finished.

The download item is added to the menu, which leads to the folder where the file was downloaded.

The navigation button has been added back and forth for navigation between pages.

implementation of the history item in process.